refactor: move Human struct and methods to store/humans.rs
